+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 ## [Unreleased]
 
+- Improve UI for adding custom commands
+- Migrate from pnpm to Bun (both extension and webviews)
+- Add Telemetry option and update dependencies
+- Refactors the extension to use webviews instead of a tree view for displaying custom commands
+
 ## [0.2.1] - 2025-01-08
 
 - Add welcome message to the custom commands tree view when no commands are added